# doc/api.yaml
openapi: 3.0.0
info:
  title: API de Gestion d'Activité
  version: 1.0.0
  description: API pour la gestion des rapports d'activité hebdomadaires, mensuels et trimestriels des départements.
security:
  - bearerAuth: []
paths:
  /login:
    post:
      summary: Authentification de l'utilisateur
      tags: [User]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Authentification réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedException'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerException'
  /signup:
    post:
      summary: Inscription de l'utilisateur
      tags: [User]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'

      responses:
        '200':
          description: Authentification réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedException'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerException'
...
  /direction:
    get:
      summary: Récupérer tous les départements
      tags: [Direction]
      responses:
        '200':
          description: Liste des départements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Direction'

    put:
      summary: Créer ou mettre à jour un département
      tags: [Direction]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string

      responses:
        '200':
          description: Département créé ou mis à jour avec succès
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedException'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerException'
  /direction/{id}:
    get:
      summary: Récupérer les informations d'un département
      tags: [Direction]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID du département
      responses:
        '200':
          description: Détails du département
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Direction'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedException'
        '404':
          description: Resource Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundException'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerException'

    delete:
      summary: Supprimer un département
      tags: [Direction]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID du département à supprimer
      responses:
        '200':
          description: Département supprimé avec succès
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedException'
        '404':
          description: Resource Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundException'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerException'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserToSite:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string

        role:
          type: string
          enum: [admin, director, employee]
        poste:
          type: string
          description: Le poste de l'utilisateur
    NewUser:
      type: object
      properties:
        id:
          type: string
        firstname:
          type: string
        lastanme:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        directionId:
          type: string
    Direction:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        responsable:
          type: array
          items:
            $ref: '#/components/schemas/UserToSite'
      required:
        - name
...
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: Resource specified Not Found
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error